Improve temperature check function, disable temporarily
This commit includes enhancements to the temperature check functionality in the health_check.py script, providing a better reporting mechanism if no temperature sensors are detected. However, this function has been temporarily disabled due to issues with its operation on macOS. It is marked as a TODO for further investigation and resolving of this issue.